app/requirements.txt
fastapi
uvicorn
debugpy
ruff
httpx
pydantic
pydantic_settings
pytest
pytest-xdist
pylint
mypy
doublex
schema
strawberry-graphql[fastapi]
pytest-asyncio
redis
types-redis
